chore: Formalize 'route walkers' a little more
